Permissions: When users are assigned an 'Author' role, they are automatically given the 'Editor' role of anything that they create.
This allows people to go back and make changes to or delete content
that they have added.

Similarly, If a user moves a piece of content from a place where they have an implicit 'Editor' role to a place where they only have an implicit 'Author' role,
they are given an explicit 'Editor' role on the newly moved content.