URLs/Migration: A URL resolver has been added that can redirect Segue 1-style urls. If the Segue 1 site has bee imported into Segue 2 then the user will be redirected to the Segue 2 version, otherwise, they will be redirected to the Segue 1 instance.
This feature allows Segue 2 to be placed at the previous location of a
Segue 1 instance (e.g. http://segue.example.edu) and the Segue 1 instance
to be moved to a new location (e.g. http://segue1.example.edu). With this
switch, Segue 2 will be able to forward old urls to either new versions
or to the Segue 1 instance for the duration of a migration period.