Login, mode-links, home, and footer all now outside page background block.
The themes used during development did not style the background-block,
so we didn't notice inconsistancies in where the login links and other
similar controls were placed.

As well, I moved the Segue-info footer to below the site rather than in
it and pushed the "Edit Site Settings" and "Permissions" buttons up into
it.