Portal speed improvements
Folders can now report a pre-sorted flag to indicate to the list that
authorization doesn't need to be checked when paginating.

Folders can also now indicate a desire not to show edit controls. This
prevents authorization traversal of huge numbers of sites when viewing
the "All Visible" Folder.

Also, added a note about slow first access to the "All Editable" folder.