Templates now import in a basic way.
Needed to disable the DomImportSiteVisitor's reqirement of a media path
as most templates won't have them. It now just throws an error if the
media path is not null or if the import tries to add a file with no
media path set.

Still need to replace #SITE_NAME# and #SITE_DESCRIPTION# placeholders.

Still need to strip out any history.