Fixed Roles issue with 'invalid default state'.
Bug 2014548
https://sourceforge.net/tracker/index.php?func=detail&aid=2014548&group_id=82171&atid=565234

It appears that since the transition to the new AuthZ2 implementation,
the roles system has not been working correctly when setting complex
authorization. This issue stems from the 'new' property of
view authorizations cascading up as well as down. This cascading up
was confusing the roles system. When there was no site-wide view
authorization the roles system thought that there was due to a view
authorization cascading up from a story or page.

This has now been sorted out and the roles system only looks at implicit
authorizations that cascade down when determining what role to display
for a user.