Improved Copy-Site status for IE and Safari.
Safari and IE do not provide access to partial data while the request is
loading, so we will just show the text "Copying Site..." and a spinning
loading icon rather than real status info for them as there is no way to
get it easily.

As well, Safari doesn't allow the setting of the Content-Length header,
for the form submission, so we will just skip that.