MediaLibrary: File library now reloads every time the media library opens.
When repeatedly adding files and making both image and text-links to them
the listings in the two panels would get out of sync. Now they are reloaded
every time the panel is opened.

There is a slight performance coast in reloading the listing, but hopefully
the ease-of-use will outweigh that.